Refactoring to consolidate common code
