+++ client/my-sites/site-settings/analytics/form-google-analytics.jsx
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import React from 'react';
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { flowRight, partialRight, pick } from 'lodash';
 
@@ -24,11 +24,75 @@ import GoogleAnalyticsSimpleForm from './form-google-analytics-simple';
  */
 import './style.scss';
 
+const validateGoogleAnalyticsCode = ( code ) =>
+	! code || code.match( /^(UA-\d+-\d+)|(G-[A-Z0-9]+)$/i );
 export const GoogleAnalyticsForm = ( props ) => {
+	const {
+		isRequestingSettings,
+		isSavingSettings,
+		enableForm,
+		translate,
+		fields,
+		updateFields,
+		trackTracksEvent,
+		eventTracker,
+		uniqueEventTracker,
+		path,
+	} = props;
+	const [ isCodeValid, setIsCodeValid ] = useState( true );
+	const [ loggedGoogleAnalyticsModified, setLoggedGoogleAnalyticsModified ] = useState( false );
+	const [ displayForm, setDisplayForm ] = useState( false );
+	const isSubmitButtonDisabled =
+		isRequestingSettings || isSavingSettings || ! isCodeValid || ! enableForm;
+	const placeholderText = isRequestingSettings ? translate( 'Loading' ) : '';
+
+	const handleFieldChange = ( key, value, callback = () => {} ) => {
+		const updatedFields = Object.assign( {}, fields.wga || {}, {
+			[ key ]: value,
+		} );
+		updateFields( { wga: updatedFields }, callback );
+	};
+
+	const handleCodeChange = ( event ) => {
+		const code = event.target.value.trim();
+		setIsCodeValid( validateGoogleAnalyticsCode( code ) );
+		handleFieldChange( 'code', code );
+	};
+
+	const recordSupportLinkClick = () => {
+		trackTracksEvent( 'calypso_traffic_settings_google_support_click' );
+	};
+
+	const handleFieldFocus = () => {
+		trackTracksEvent( 'calypso_google_analytics_key_field_focused', { path } );
+		eventTracker( 'Focused Analytics Key Field' )();
+	};
+
+	const handleFieldKeypress = () => {
+		if ( ! loggedGoogleAnalyticsModified ) {
+			trackTracksEvent( 'calypso_google_analytics_key_field_modified', { path } );
+			setLoggedGoogleAnalyticsModified( true );
+		}
+		uniqueEventTracker( 'Typed In Analytics Key Field' )();
+	};
+
+	const newProps = {
+		...props,
+		displayForm,
+		handleCodeChange,
+		handleFieldChange,
+		handleFieldFocus,
+		handleFieldKeypress,
+		isCodeValid,
+		isSubmitButtonDisabled,
+		placeholderText,
+		recordSupportLinkClick,
+		setDisplayForm,
+	};
 	if ( props.siteIsJetpack ) {
-		return <GoogleAnalyticsJetpackForm { ...props } />;
+		return <GoogleAnalyticsJetpackForm { ...newProps } />;
 	}
-	return <GoogleAnalyticsSimpleForm { ...props } />;
+	return <GoogleAnalyticsSimpleForm { ...newProps } />;
 };
 
 const mapStateToProps = ( state ) => {
